% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_api_wrapper.R
\name{create_wrapper_function}
\alias{create_wrapper_function}
\title{Create Wrapper Function}
\usage{
create_wrapper_function(api, operations, op_name)
}
\arguments{
\item{api}{The API object returned by rapiclient::get_api()}

\item{op_name}{The name of the operation}
}
\value{
A wrapper function for the specified operation
}
\description{
This function creates a wrapper function for a specific API operation.
}
