% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_api_wrapper.R
\name{generate_api_wrappers}
\alias{generate_api_wrappers}
\title{Generate API Wrapper Functions}
\usage{
generate_api_wrappers(api, operations)
}
\arguments{
\item{api}{The API object returned by rapiclient::get_api()}
}
\value{
A list of generated wrapper functions
}
\description{
This function dynamically generates wrapper functions for all operations in the API.
}
