% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_api_wrapper.R
\name{initialize_api_with_wrappers}
\alias{initialize_api_with_wrappers}
\title{Initialize API and Generate Wrappers}
\usage{
initialize_api_with_wrappers(url, config = NULL, type = NULL)
}
\arguments{
\item{url}{The URL of the API}
}
\value{
A list containing the API object and generated wrapper functions
}
\description{
This function initializes the API and generates wrapper functions for all operations.
}
