% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_api_wrapper.R
\name{list_api_operations}
\alias{list_api_operations}
\title{List Available API Operations}
\usage{
list_api_operations(api_wrappers)
}
\arguments{
\item{api_wrappers}{The list of API wrappers returned by initialize_api_with_wrappers()}
}
\value{
A data frame of operation names and descriptions
}
\description{
This function lists all available API operations and their basic information.
}
